[bir][gcir] Separate Passes.td for conversions and transforms
