+++ include/belalang/BIR/Transforms/Passes.h
@@ -1,5 +1,5 @@
-#ifndef BELALANG_PASSES_H_
-#define BELALANG_PASSES_H_
+#ifndef BELALANG_TRANSFORMS_PASSES_H_
+#define BELALANG_TRANSFORMS_PASSES_H_
 
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/Dialect/ControlFlow/IR/ControlFlow.h"
@@ -15,17 +15,15 @@ namespace belalang {
 namespace bir {
 
 #define GEN_PASS_DECL
-#include "belalang/BIR/Passes.h.inc"
+#include "belalang/BIR/Transforms/Passes.h.inc"
 
 #define GEN_PASS_REGISTRATION
-#include "belalang/BIR/Passes.h.inc"
+#include "belalang/BIR/Transforms/Passes.h.inc"
 
 void populateBelalangFlattenCFGPatterns(mlir::RewritePatternSet &patterns);
 void populateBelalangLowerDeclToMemoryPatterns(
     mlir::RewritePatternSet &patterns, mlir::DataFlowSolver &solver);
 void populateBelalangLowerFuncExprPatterns(mlir::RewritePatternSet &patterns);
-void populateBelalangBIRToLLVMPatterns(mlir::RewritePatternSet &patterns,
-                                       mlir::TypeConverter &typeConverter);
 
 // -----------------------------------------------------------------------------
 // Pipelines
@@ -57,4 +55,4 @@ void registerBIRPipelines();
 } // namespace bir
 } // namespace belalang
 
-#endif // BELALANG_PASSES_H_
+#endif // BELALANG_TRANSFORMS_PASSES_H_
