+++ include/belalang/BIR/Transforms/Passes.td
@@ -1,5 +1,5 @@
-#ifndef BELALANG_PASSES_TD_
-#define BELALANG_PASSES_TD_
+#ifndef BELALANG_TRANSFORMS_PASSES_TD_
+#define BELALANG_TRANSFORMS_PASSES_TD_
 
 include "mlir/Pass/PassBase.td"
 
@@ -36,11 +36,4 @@ def BelalangOptimizeStructLayoutPass : Pass<"bir-optimize-struct-layout"> {
                            "::mlir::LLVM::LLVMDialect"];
 }
 
-def BelalangBIRToLLVMPass : Pass<"convert-bir-to-llvm"> {
-  let summary = "lower bir instructions to LLVM";
-  let dependentDialects = ["::belalang::bir::BIRDialect",
-                           "::mlir::cf::ControlFlowDialect",
-                           "::mlir::LLVM::LLVMDialect"];
-}
-
-#endif // BELALANG_PASSES_TD_
+#endif // BELALANG_TRANSFORMS_PASSES_TD_
