--- include/belalang/GCIR/Transforms/CMakeLists.txt
@@ -1,3 +1,0 @@
-set(LLVM_TARGET_DEFINITIONS Passes.td)
-mlir_tablegen(Passes.h.inc -gen-pass-decls)
-add_public_tablegen_target(BelalangGCIRPassIncGen)
