+++ lib/BIR/CMakeLists.txt
@@ -2,7 +2,8 @@ add_library(BelalangBIR)
 add_dependencies(BelalangBIR
   BelalangBIRIncGen
   BelalangBIRLoopOpInterfaceIncGen
-  BelalangBIRPassIncGen
+  BelalangBIRTransformPassIncGen
+  BelalangBIRConversionPassIncGen
 )
 target_link_libraries(BelalangBIR PUBLIC
   MLIRAnalysis
@@ -22,3 +23,4 @@ add_subdirectory(Analysis)
 add_subdirectory(IR)
 add_subdirectory(Interfaces)
 add_subdirectory(Transforms)
+add_subdirectory(Conversions)
