+++ lib/BIR/Conversions/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(BelalangBIR PRIVATE
+  BIRToLLVM.cpp
+)
