+++ lib/BIR/Transforms/VerifyLoweredForm.cpp
@@ -1,9 +1,9 @@
 #include "belalang/BIR/IR/BIR.h"
-#include "belalang/BIR/Passes.h"
+#include "belalang/BIR/Transforms/Passes.h"
 
 namespace mlir {
 #define GEN_PASS_DEF_BELALANGVERIFYLOWEREDFORMPASS
-#include "belalang/BIR/Passes.h.inc"
+#include "belalang/BIR/Transforms/Passes.h.inc"
 } // namespace mlir
 
 namespace {
