+++ lib/GCIR/Conversions/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(BelalangGCIR PRIVATE
+  GCIRToLLVM.cpp
+)
